Move `dict_to_array` to helpers
